% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geomaps.R
\name{Geomaps}
\alias{Geomaps}
\title{Grafica Mapas Con las caracteristicas de los suelos, usando la api de google maps, y una data dada}
\usage{
Geomaps(ApiKey, FileRoute, Elemento, Rango1, Rango2, Rango3, Rango4)
}
\arguments{
\item{ApiKey}{Kei of GoogleCloud}

\item{FileRoute}{Route File}

\item{Elemento}{Patron que se quiere graficar}

\item{Rango}{4 rangos.}
}
\description{
Grafica Mapas Con las caracteristicas de los suelos, usando la api de google maps, y una data dada
}
