% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package1.R
\docType{package}
\name{package1}
\alias{package1}
\alias{package1-package}
\title{DBUmaps Package}
\description{
Contiene Funcion que grafica mapas dinamicos con las caracteristicas de los suelos
}
\author{
Diego Bolivar \email{diego.bolivar@unet.edu.ve}
}
